1. ReactNative要素查询模块：空间查询测试 2. 修订要素空间属性查询，一直查询不到数据的问题 3. 新建ReactNative测试程序：测试图形覆盖物模块功能组件接口 4. 修订添加地图覆盖物，程序一直崩溃的问题 5. 修订在地图上添加GraphicImage始终看不到当前图片的问题

+++ android/app/src/main/java/com/mapgis_mobile_reactnative/utils/ConvertUtil.java
@@ -46,10 +46,10 @@ public class ConvertUtil
 				String strs[] = strs1.split(",");
 				if (strs != null && strs.length == 4)
 				{
-					red = Integer.parseInt(strs[0]);
-					green = Integer.parseInt(strs[1]);
-					blue = Integer.parseInt(strs[2]);
-					alpha = Integer.parseInt(strs[3]);
+					red = (int) Double.parseDouble(strs[0]);
+					green = (int) Double.parseDouble(strs[1]);
+					blue = (int) Double.parseDouble(strs[2]);
+					alpha = (int) Double.parseDouble(strs[3]);
 				}
 			}
 		}
